requirements.txt
django == 2.1
nltk == 3.3
gensim == 3.5.0
six == 1.11.0
numpy == 1.15.0
docx2txt == 0.7
pdfminer3k == 1.3.1
textract == 1.6.1